Work in progress game server and UI
Implements an initial take on a Clojure game server with web sockets and
simple hexagon grid generation. Uses simplex noise to generate random
terrain. Uses the cube/axial coordinate system, and does some coordinate
mapping (from hex to cartesian coordinates) to facilitate mapping to the
Simplex noise coordinate space. The implementation will probably need
fine tuning.

Implements a UI that render the game using Vite.js, React, and Three.js
Fiber. This allows us to render the game in 3D with simple shapes and
some illumintation settings and an orbital camera. The frontend will use
a web socket to receive game state updates from the server, though the
scaffolding for this is not yet in place.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>