Add web socket comms and improved graphics and new models
Initial implementation of some improved graphics. Includes a skybox,
textures for the hexagons, and a spaceship model to represent players.
Spaceship model and textures are licensed under CC licenses and are
redistributable, thus being a great fit for an open source project.
The skybox was generated using a procedural skybox generator which
should mean it is okay to use the result in this project.

Added a web socket library and code to receive messages from the game
server. Currently the client can receive game state messages from the
server telling the client what the current game state is. This allows
the client to draw out the hexagon map that the server generated.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>