Use zustand for state and refactor components
Initial work moving all the state into a separate state handling
library. Zustand is built by the same author as Three Fiber and seems
well thought out and performant enough for this project.

It seems to be loosely inspired by Redux and other state management
libraries, but is slightly less opinionated and allows you some
additional freedom. Further more it is a bit simpler to use it
performantly with Three Fiber.

Initially the game state is set up with empty values, i.e. empty arrays
for players, spaceships, and hex tiles. When server state is received
from the web socket we initialize the state store. The Graphics
component uses useStore to get the game state and uses it to render the
hexagons and spaceships.

The hexagon function to render hexagons were refactored into a separate
module. For both spaceships and hexagons we now use THREE.Vector3
objects in our state to easily render them with the correct position.

Todo: Figure out the best way to animate the height of the spaceships to
create a "hover" effect.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>