Use spring for spaceship movement and rotation animation
The game state now keeps track of the spaceships' position (in both
cartesian and hexagonal sense) and the spaceships' rotation. The
Spaceship component has local state to keep track of the previous and
next positions. This makes it possible to use the React Spring library
(for Three Fiber) to animate the movement and rotation of the ship.

Currently rotation and movement is implemented as two separate springs,
the rotation being done immediately and the movement delayed to run
after rotation is complete. The springs use an easing physics preset,
but it is possible to adjust the settings if needed.

Server side code has been added to find random neighbor tiles and
moving the spaceships around to see the movement animation in the
frontend in action. Eventually the player AIs will be responsible for
moving the spaceships.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>