feat: AI game state loop and improved state handling
Introduces new background loop for running the game. The loop works by
sending the game state to each AI awaiting their reply. Upon reply it
checks which actions the AIs performed (if any) and update the game
state correspondingly. Currently this check is work in progress.

The loop is wrapped in a future to run the loop in a separate thread.
This allows the main thread to continue other work while the game loop
operates.

A new game-state flag atom is introduced to control the running of the
game. It starts off as :wait and only when changed to :play will the
game begin. This is done using the add-watch function.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>