Implement better collision animations
Handle collisions by adding collision coordinates (cartesian)
and event type to spaceships state and use conditionals in the use
effects calls to determine how to update the location state of the
spaceship.

This works out nicely to animate the spaceship moving halfway to the
intended coordinate before "colliding" and backing up to its original
position.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>