Make Dockerfile and build setup
Adds tools.build setup for building an uberjar of the Clojure server.
This makes is easy to build and then run the server. A new core
namespace is used to provide the main method and start the server when
running as a jar.

A ring file handler is used to serve the frontend. The frontend is
served on the root path of the server. The socket has been moved to the
/socket path.

The frontend has been amended and public assets (textures, models, etc)
been moved to the public folder which is automatically bundled/included
by Vite in the production build.

A multi stage Docker container is provided to make it easy to build the
project and run multiple servers concurrently by spinning up multiple
containers.

The frontend will connect to the hardcoded localhost:8080/socket url in
dev. In production it will connect to {window.host}/socket which works
as the UI is now hosted on the same server running the socket.